% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importsurvey-package.R
\docType{package}
\name{importsurvey-options}
\alias{importsurvey-options}
\title{Package options}
\description{
Run \code{\link[=show_options]{show_options()}} to see available options.
}
\seealso{
Other options: 
\code{\link{show_options}()}
}
\concept{options}
