% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importsurvey-package.R
\docType{package}
\name{importsurvey-package}
\alias{importsurvey}
\alias{importsurvey-package}
\title{importsurvey: Import data into R, with an application to the National Center for Health Statistics (NCHS)}
\description{
Import data into R, with an application to the National Center for Health Statistics (NCHS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cdcgov.github.io/importsurvey}
  \item \url{https://github.com/CDCgov/importsurvey}
}

}
\author{
\strong{Maintainer}: Alex Strashny \email{AStrashny@cdc.gov}

}
\keyword{internal}
