% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_options.R
\name{show_options}
\alias{show_options}
\title{Show package options}
\usage{
show_options(sw = "importsurvey")
}
\arguments{
\item{sw}{starting characters}
}
\value{
List of options and their values.
}
\description{
Show package options
}
\examples{
show_options()
}
\seealso{
Other options: 
\code{\link{importsurvey-options}}
}
\concept{options}
