% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sas.R
\name{import_sas}
\alias{import_sas}
\title{Import SAS data into R}
\usage{
import_sas(sas_data, sas_formats_data, formats = "attr")
}
\arguments{
\item{sas_data}{SAS survey data file}

\item{sas_formats_data}{SAS formats data file (produced with CNTLOUT option of PROC FORMAT)}

\item{formats}{how are formats specified?

\code{formats}: how are formats specified?
\itemize{
\item "attr": in \verb{attr(*, "format.sas")}
\item "name": same as the variable name
}}
}
\value{
\code{data.frame}. If this is a complex survey, use \code{svydesign} to
create a survey design object.
}
\description{
Import SAS data into R
}
\examples{
\dontrun{
d1 = import_sas(sas_data, sas_formats_data)
}
}
\seealso{
Other import: 
\code{\link{import_sas_namcs2019puf}()},
\code{\link{import_sas_rcc2018puf}()}
}
\concept{import}
